gradebotguru/grader.py
from typing import Dict, Tuple, Any
from gradebotguru.prompts import generate_prompt
from gradebotguru.response_parser import parse_response
from gradebotguru.llm_interface.base_llm import BaseLLM


def preprocess_submission(submission: str) -> str:
    """
    Preprocess the submission before sending it to the LLM.

    Parameters:
    - submission (str): The original submission text.

    Returns:
    - str: The preprocessed submission text.
    """
    # Placeholder for preprocessing logic
    return submission


def postprocess_results(grade: float, feedback: str) -> Tuple[float, str]:
    """
    Postprocess the results obtained from the LLM.

    Parameters:
    - grade (float): The grade received from the LLM.
    - feedback (str): The feedback received from the LLM.

    Returns:
    - tuple: The postprocessed grade and feedback.
    """
    # Placeholder for postprocessing logic
    return grade, feedback


def grade_submission(submission: str, rubric: Dict[str, Dict[str, Any]], llm_interface: BaseLLM) -> Tuple[float, str]:
    """
    Grade a single submission based on the provided rubric.

    Parameters:
    - submission (str): The student's submission.
    - rubric (dict): The grading rubric.
    - llm_interface (BaseLLM): The LLM interface to generate responses.

    Returns:
    - tuple: The grade and feedback for the submission.
    """
    preprocessed_submission = preprocess_submission(submission)
    prompt = generate_prompt(rubric, preprocessed_submission)
    response = llm_interface.get_response(prompt)
    result = parse_response(response)
    postprocessed_grade, postprocessed_feedback = postprocess_results(result['grade'], result['feedback'])
    return postprocessed_grade, postprocessed_feedback
